chore: 상수데이터 분리
