--- week3_cardgame/cardgame/src/components/card/CardData.jsx
@@ -1,22 +1,0 @@
-import card1 from './../../assets/images/adora.png';
-import card2 from './../../assets/images/candlehead.png';
-import card3 from './../../assets/images/citrusella.png';
-import card4 from './../../assets/images/crumbelina.png';
-import card5 from './../../assets/images/vanellope.png';
-import card6 from './../../assets/images/jubileena.png';
-import card7 from './../../assets/images/rancis.png';
-import card8 from './../../assets/images/snowanna.png';
-import card9 from './../../assets/images/taffyta.png';
-
-const CARDLIST = [
-    { id: 1, img: card1 },
-    { id: 2, img: card2 },
-    { id: 3, img: card3 },
-    { id: 4, img: card4 },
-    { id: 5, img: card5 },
-    { id: 6, img: card6 },
-    { id: 7, img: card7 },
-    { id: 8, img: card8 },
-    { id: 9, img: card9 },
-];
-export default CARDLIST;
