+++ week3_cardgame/cardgame/src/constants/Level.jsx
@@ -1,0 +1,5 @@
+export const GAME_LEVELS = [
+    { label: 'Easy', level: 5 },
+    { label: 'Normal', level: 7 },
+    { label: 'Hard', level: 9 },
+];
